test(loader): add tests for non-existing project root
This tests are run in OS temp directory as they need to be run in place where there is no
`package.json` in the whole filesystem tree. `/tmp` seems like a good place because there is no
sensible reason to have `package.json` in `/tmp` or `/`. If you have `package.json` there, please
tell me WHY.