test(esmlm): restore coverage hack
It seems that the culprit is the nesting of processes: ava → `execa` in tests → `execa` in bin. If
the coverage is disabled directly in the bin's `execa`, the coverage calculation works correctly.
However, bin should not change envs at all therefore for now the only sensible solution is to
disable the coveraga for all `execa`s.